Тест-2

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -42,6 +42,9 @@ public class UserService {
     }
 
     public void deleteUser(int id) {
+        if (id < 0) {
+            throw new ObjectNotFoundException(LogMessagesUsers.ID_NOT_POSITIVE.getMessage());
+        }
         userStorage.delete(id);
     }
